Allow more flexible lists of lights in scenes
The list of lights in a scene may now be a subset of the lights in its
parent group. Calling 'Modify scene' with a light state for a light not
in the scene, but part of the scene's parent group, will add the light
and its state to the scene. Calling it with an empty light state for a
light in the scene will remove that light from the scene.