Add scaffolding for 'setHueLightState()'
'setLightState()' calls out to this new helper method when dealing
with a Hue light that supports effects with a request that contains
only items that can be handled by the manufacturer-specific command.

The new method will eventually build a '0x00' command from the
'0xfc03' cluster to control the light, mimicking what the Hue Bridge.