Add rudimentary recall of Hue Dynamic Scenes
Playing a Hue Dynamic Scene first recalls a “regular” scene and then
instructs the members of that scene to begin playing the ‘dynamic
palette’. The effect works best when the recalled scene puts its
member lights into a state that is identical to one of those provided
in the ‘dynamic palette’. The light’s firmware takes care of playing
the scene starting from the lights current state.

The equivalent of the contents of the ‘palette’ object in the CLIP v2
API must be provided to the ‘recall’ endpoint of ‘hue-scenes’ every
time - it is not stored by the API. Currently only a payload of colors
is supported.