Add Dynamic Scene info to other parts of the API
Scene Attributes will now include the definitions of dynamic scenes for
scenes that have them.

Recalling a scene that also has a dynamic scene with the ‘auto_dynamic’
attribute will play the dynamic scene immediately after recalling the
scene.

Playing a dynamic scene without providing a definition for it will
attempt to play a stored definition, if one is available.