Ensure scenes without a dynamic scene are not affected
Trying to fetch JSON strings for DynamicSceneStates for Scenes that
don’t have one was causing issues.

DynamicSceneStates won’t be created for scenes that don’t have one when
reading them off the database.

Allow deleting the defition of a dynamic scene by passing an empty body
to the REST call.